ck-builder: ckt::run support for ck-tile bwd weight
