+++ profiler/include/profiler/grouped_convolution_forward_tile_algs.hpp
@@ -9,7 +9,7 @@
 #include "grouped_convolution_signatures.hpp"
 
 #include "ck_tile/builder/testing/filter_extent.hpp"
-#include "ck_tile/builder/testing/conv/fwd_ck_tile.hpp"
+#include "ck_tile/builder/testing/conv/ck_tile.hpp"
 #include "ck_tile/builder/testing/conv/reference.hpp"
 #include "ck_tile/builder/conv_builder.hpp"
